+++ wai-extra/test/WaiExtraSpec.hs
@@ -19,15 +19,28 @@ import Data.Monoid (mempty, mappend)
 import qualified Data.Text as TS
 import qualified Data.Text.Encoding as TE
 import qualified Data.Text.Lazy as T
-import Network.HTTP.Types (status200)
+import Network.HTTP.Types (
+    Header,
+    RequestHeaders,
+    ResponseHeaders,
+    hContentEncoding,
+    hContentLength,
+    hContentType,
+    partialContent206,
+    status200,
+ )
+import Network.HTTP.Types.Header (hAcceptEncoding, hVary)
 import Network.Wai
+import System.Directory (listDirectory)
+import System.IO.Temp (withSystemTempDirectory)
 import System.Log.FastLogger (fromLogStr)
-import Test.HUnit (Assertion, (@?=))
+import Test.HUnit (Assertion, assertBool, assertEqual, (@?=))
 import Test.Hspec
 
+import Network.Wai.Header (parseQValueList)
 import Network.Wai.Middleware.AcceptOverride (acceptOverride)
 import Network.Wai.Middleware.Autohead (autohead)
-import Network.Wai.Middleware.Gzip (def, defaultCheckMime, gzip)
+import Network.Wai.Middleware.Gzip (GzipFiles (..), GzipSettings (..), def, defaultCheckMime, gzip)
 import Network.Wai.Middleware.Jsonp (jsonp)
 import Network.Wai.Middleware.MethodOverride (methodOverride)
 import Network.Wai.Middleware.MethodOverridePost (methodOverridePost)
@@ -51,10 +64,15 @@ spec = do
     , it "takeLine" caseTakeLine
     -}
     it "jsonp" caseJsonp
-    it "gzip" caseGzip
-    it "gzip not for MSIE" caseGzipMSIE
-    it "gzip bypass when precompressed" caseGzipBypassPre
-    it "defaultCheckMime" caseDefaultCheckMime
+    describe "gzip" $ do
+        it "gzip" caseGzip
+        it "more gzip" caseGzip2
+        it "gzip not on partial content" caseGzipPartial
+        it "gzip removes length header" caseGzipLength
+        it "gzip not for MSIE" caseGzipMSIE
+        it "gzip bypass when precompressed" caseGzipBypassPre
+        it "defaultCheckMime" caseDefaultCheckMime
+        it "gzip checking of files" caseGzipFiles
     it "vhost" caseVhost
     it "autohead" caseAutohead
     it "method override" caseMethodOverride
@@ -65,6 +83,7 @@ spec = do
     it "stream LBS" caseStreamLBS
     it "can modify POST params before logging" caseModifyPostParamsInLogs
     it "can filter requests in logs" caseFilterRequestsInLogs
+    it "can parse Q values" caseQValues
 
 toRequest :: S8.ByteString -> S8.ByteString -> SRequest
 toRequest ctype content = SRequest defaultRequest
@@ -133,7 +152,7 @@ jsonpApp = jsonp $ \_ f -> f $ responseLBS
     "{\"foo\":\"bar\"}"
 
 caseJsonp :: Assertion
-caseJsonp = flip runSession jsonpApp $ do
+caseJsonp = withSession jsonpApp $ do
     sres1 <- request defaultRequest
                 { queryString = [("callback", Just "test")]
                 , requestHeaders = [("Accept", "text/javascript")]
@@ -156,33 +175,199 @@ caseJsonp = flip runSession jsonpApp $ do
     assertBody "{\"foo\":\"bar\"}" sres3
 
 gzipApp :: Application
-gzipApp = gzip def $ \_ f -> f $ responseLBS status200
-    [("Content-Type", "text/plain")]
-    "test"
+gzipApp = gzipApp' id
 
--- Lie a little and don't compress the body.  This way we test
--- that the compression is skipped based on the presence of
--- the Content-Encoding header.
-gzipPrecompressedApp :: Application
-gzipPrecompressedApp = gzip def $ \_ f -> f $ responseLBS status200
-    [("Content-Type", "text/plain"), ("Content-Encoding", "gzip")]
-    "test"
+gzipApp' :: (Response -> Response) -> Application
+gzipApp' changeRes =
+    gzip def $ \_ f -> f . changeRes $ responseLBS status200
+        [("Content-Type", "text/plain")]
+        "test"
+
+gzipAppWithHeaders :: ResponseHeaders -> Application
+gzipAppWithHeaders hdrs = gzipApp' $ mapResponseHeaders $ (hdrs ++)
+
+gzipFileApp :: GzipSettings -> Application
+gzipFileApp = flip gzipFileApp' id
+
+gzipJSONFile, gzipNoPreCompressFile :: FilePath
+gzipJSONFile = "test/json"
+gzipNoPreCompressFile = "test/noprecompress"
+
+gzipJSONBody, gzipNocompressBody :: L.ByteString
+#if WINDOWS
+gzipJSONBody = "{\"data\":\"this is some data\"}\r\n"
+gzipNocompressBody = "noprecompress\r\n"
+#else
+gzipJSONBody = "{\"data\":\"this is some data\"}\n"
+gzipNocompressBody = "noprecompress\n"
+#endif
+
+-- | Use 'changeRes' to make r
+gzipFileApp' :: GzipSettings -> (Response -> Response) -> Application
+gzipFileApp' set changeRes =
+    gzip set $ \_ f -> f . changeRes $
+        responseFile status200 [(hContentType, "application/json")] gzipJSONFile Nothing
+
+acceptGzip :: Header
+acceptGzip = (hAcceptEncoding, "gzip")
+
+doesEncodeGzip :: RequestHeaders -> Session SResponse
+doesEncodeGzip = doesEncodeGzip' "test"
+
+doesEncodeGzipJSON :: RequestHeaders -> Session SResponse
+doesEncodeGzipJSON = doesEncodeGzip' gzipJSONBody
+
+doesEncodeGzipNoPreCompress :: RequestHeaders -> Session SResponse
+doesEncodeGzipNoPreCompress = doesEncodeGzip' gzipNocompressBody
+
+doesEncodeGzip' :: L.ByteString -> RequestHeaders -> Session SResponse
+doesEncodeGzip' body hdrs = do
+    sres <- request defaultRequest
+        { requestHeaders = hdrs
+        }
+    assertHeader hContentEncoding "gzip" sres
+    assertHeader hVary "Accept-Encoding" sres
+    liftIO $ decompress (simpleBody sres) @?= body
+    pure sres
+
+doesNotEncodeGzip :: RequestHeaders -> Session SResponse
+doesNotEncodeGzip = doesNotEncodeGzip' "test"
+
+doesNotEncodeGzipJSON :: RequestHeaders -> Session SResponse
+doesNotEncodeGzipJSON = doesNotEncodeGzip' gzipJSONBody
+
+doesNotEncodeGzipNoPreCompress :: RequestHeaders -> Session SResponse
+doesNotEncodeGzipNoPreCompress = doesNotEncodeGzip' gzipNocompressBody
+
+doesNotEncodeGzip' :: L.ByteString -> RequestHeaders -> Session SResponse
+doesNotEncodeGzip' body hdrs = do
+    sres <- request defaultRequest
+        { requestHeaders = hdrs
+        }
+    assertNoHeader hContentEncoding sres
+    assertHeader hVary "Accept-Encoding" sres
+    assertBody body sres
+    pure sres
 
 caseGzip :: Assertion
-caseGzip = flip runSession gzipApp $ do
-    sres1 <- request defaultRequest
-                { requestHeaders = [("Accept-Encoding", "gzip")]
-                }
-    assertHeader "Content-Encoding" "gzip" sres1
-    assertHeader "Vary" "Accept-Encoding" sres1
-    liftIO $ decompress (simpleBody sres1) @?= "test"
+caseGzip = do
+    withSession gzipApp $ do
+        _ <- doesEncodeGzip [acceptGzip]
+        _ <- doesNotEncodeGzip []
+        _ <- doesEncodeGzip [(hAcceptEncoding, "compress , gzip ; q=0.8")]
+        pure ()
 
-    sres2 <- request defaultRequest
-                { requestHeaders = []
-                }
-    assertNoHeader "Content-Encoding" sres2
-    assertHeader "Vary" "Accept-Encoding" sres2
-    assertBody "test" sres2
+    withSession (gzipAppWithHeaders [(hContentLength, "200")]) $ do
+        sres4 <- doesNotEncodeGzip [acceptGzip]
+        assertHeader hContentLength "200" sres4
+
+caseGzipLength :: Assertion
+caseGzipLength = do
+    withSession (gzipAppWithHeaders [(hContentLength, "4000")]) $ do
+        sres <- doesEncodeGzip [acceptGzip]
+        assertNoHeader hContentLength sres
+
+caseGzipPartial :: Assertion
+caseGzipPartial =
+    withSession partialApp $ do
+        _ <- doesNotEncodeGzip [acceptGzip]
+        pure ()
+  where
+    partialApp = gzipApp' $ mapResponseStatus $ const partialContent206
+
+-- | Checking that it doesn't compress when already compressed AND
+-- doesn't replace already set "Vary" header.
+caseGzip2 :: Assertion
+caseGzip2 =
+    withSession gzipVariantApp $ do
+        sres1 <- request defaultRequest
+                    { requestHeaders = [(hAcceptEncoding, "compress, gzip")] }
+        assertHeader hContentEncoding "compress" sres1
+        assertHeader hVary "Accept-Encoding, foobar" sres1
+  where
+    gzipVariantApp = gzipAppWithHeaders
+        [ ("Content-Encoding", "compress")
+        , ("Vary", "foobar")
+        ]
+
+-- | Testing of the GzipSettings's 'GzipFiles' setting
+-- with 'ResponseFile' responses.
+caseGzipFiles :: Assertion
+caseGzipFiles = do
+    -- Default GzipSettings ignore compressing files
+    withSession (gzipFileApp def) $ do
+        _ <- doesNotEncodeGzipJSON [acceptGzip]
+        _ <- doesNotEncodeGzipJSON []
+        pure ()
+
+    -- Just compresses the file
+    withSession (gzipFileApp def{gzipFiles = GzipCompress}) $ do
+        _ <- doesEncodeGzipJSON [acceptGzip]
+        _ <- doesNotEncodeGzipJSON []
+        pure ()
+
+    -- Checks for a "filename.gz" file in the same folder
+    withSession (gzipFileApp def{gzipFiles = GzipPreCompressed GzipIgnore}) $ do
+        sres <- request defaultRequest
+            { requestHeaders = [acceptGzip] }
+        assertHeader hContentEncoding "gzip" sres
+        assertHeader hVary "Accept-Encoding" sres
+        -- json.gz has body "test\n"
+        assertBody
+#if WINDOWS
+            "test\r\n"
+#else
+            "test\n"
+#endif
+            sres
+
+        doesNotEncodeGzipJSON [] >> pure ()
+
+    -- If no "filename.gz" file is in the same folder, just ignore
+    withSession (noPreCompressApp $ GzipPreCompressed GzipIgnore) $ do
+        _ <- doesNotEncodeGzipNoPreCompress [acceptGzip]
+        _ <- doesNotEncodeGzipNoPreCompress []
+        pure ()
+
+    -- If no "filename.gz" file is in the same folder, just compress
+    withSession (noPreCompressApp $ GzipPreCompressed GzipCompress) $ do
+        _ <- doesEncodeGzipNoPreCompress [acceptGzip]
+        _ <- doesNotEncodeGzipNoPreCompress []
+        pure ()
+
+    -- Using a caching directory
+    withSystemTempDirectory "gziptest" $ \path -> do
+        let checkTempDir n s = do
+                fs <- listDirectory path
+                assertBool s $ length fs == n
+        checkTempDir 0 "temp directory should be empty"
+        -- Respond with "test/json" file
+        withSession (gzipFileApp def{gzipFiles = GzipCacheFolder path}) $ do
+            _ <- doesEncodeGzipJSON [acceptGzip]
+            liftIO $ checkTempDir 1 "should have one file"
+            _ <- doesEncodeGzipJSON [acceptGzip]
+            liftIO $ checkTempDir 1 "should still have only one file"
+            _ <- doesNotEncodeGzipJSON []
+            liftIO $ checkTempDir 1 "should not have done anything"
+
+        -- Respond with "test/noprecompress" file
+        withSession (noPreCompressApp $ GzipCacheFolder path) $ do
+            _ <- doesEncodeGzipNoPreCompress [acceptGzip]
+            liftIO $ checkTempDir 2 "should now have 2 files"
+            _ <- doesEncodeGzipNoPreCompress [acceptGzip]
+            liftIO $ checkTempDir 2 "should still only have 2 files"
+            _ <- doesNotEncodeGzipNoPreCompress []
+            liftIO $ checkTempDir 2 "again should not have done anything"
+
+        -- try "test/json" again, just to make sure it isn't a weird Session bug
+        withSession (gzipFileApp def{gzipFiles = GzipCacheFolder path}) $ do
+            _ <- doesEncodeGzipJSON [acceptGzip]
+            liftIO $ checkTempDir 2 "just to make sure it isn't a fluke"
+
+  where
+    noPreCompressApp set = gzipFileApp'
+        def{gzipFiles = set}
+        $ const $ responseFile status200 [(hContentType, "text/plain")] gzipNoPreCompressFile Nothing
 
 caseDefaultCheckMime :: Assertion
 caseDefaultCheckMime = do
@@ -195,25 +380,23 @@ caseDefaultCheckMime = do
     go "application/json; charset=utf-8" True
 
 caseGzipMSIE :: Assertion
-caseGzipMSIE = flip runSession gzipApp $ do
-    sres1 <- request defaultRequest
-                { requestHeaders =
-                    [ ("Accept-Encoding", "gzip")
-                    , ("User-Agent", "Mozilla/4.0 (Windows; MSIE 6.0; Windows NT 6.0)")
-                    ]
-                }
-    assertNoHeader "Content-Encoding" sres1
+caseGzipMSIE = withSession gzipApp $ do
+    sres1 <- doesNotEncodeGzip
+        [ acceptGzip
+        , ("User-Agent", "Mozilla/4.0 (Windows; MSIE 6.0; Windows NT 6.0)")
+        ]
     assertHeader "Vary" "Accept-Encoding" sres1
-    liftIO $ simpleBody sres1 @?= "test"
 
 caseGzipBypassPre :: Assertion
-caseGzipBypassPre = flip runSession gzipPrecompressedApp $ do
-    sres1 <- request defaultRequest
-                { requestHeaders = [("Accept-Encoding", "gzip")]
-                }
-    assertHeader "Content-Encoding" "gzip" sres1
-    assertHeader "Vary" "Accept-Encoding" sres1
-    assertBody "test" sres1 -- the body is not actually compressed
+caseGzipBypassPre =
+    -- Lie a little and don't compress the body.  This way we test
+    -- that the compression is skipped based on the presence of
+    -- the Content-Encoding header.
+    withSession (gzipAppWithHeaders [(hContentEncoding, "gzip")]) $ do
+        sres1 <- request defaultRequest{ requestHeaders = [acceptGzip] }
+        assertHeader "Content-Encoding" "gzip" sres1
+        assertHeader "Vary" "Accept-Encoding" sres1
+        assertBody "test" sres1 -- the body is not actually compressed
 
 vhostApp1, vhostApp2, vhostApp :: Application
 vhostApp1 _ f = f $ responseLBS status200 [] "app1"
@@ -224,7 +407,7 @@ vhostApp = vhost
     vhostApp2
 
 caseVhost :: Assertion
-caseVhost = flip runSession vhostApp $ do
+caseVhost = withSession vhostApp $ do
     sres1 <- request defaultRequest
                 { requestHeaders = [("Host", "foo.com")]
                 }
@@ -240,7 +423,7 @@ autoheadApp = autohead $ \_ f -> f $ responseLBS status200
     [("Foo", "Bar")] "body"
 
 caseAutohead :: Assertion
-caseAutohead = flip runSession autoheadApp $ do
+caseAutohead = withSession autoheadApp $ do
     sres1 <- request defaultRequest
                 { requestMethod = "GET"
                 }
@@ -258,7 +441,7 @@ moApp = methodOverride $ \req f -> f $ responseLBS status200
     [("Method", requestMethod req)] ""
 
 caseMethodOverride :: Assertion
-caseMethodOverride = flip runSession moApp $ do
+caseMethodOverride = withSession moApp $ do
     sres1 <- request defaultRequest
                 { requestMethod = "GET"
                 , queryString = []
@@ -281,7 +464,7 @@ mopApp :: Application
 mopApp = methodOverridePost $ \req f -> f $ responseLBS status200 [("Method", requestMethod req)] ""
 
 caseMethodOverridePost :: Assertion
-caseMethodOverridePost = flip runSession mopApp $ do
+caseMethodOverridePost = withSession mopApp $ do
 
     -- Get Request are unmodified
     sres1 <- let r = toRequest "application/x-www-form-urlencoded" "_method=PUT&foo=bar&baz=bin"
@@ -308,7 +491,7 @@ aoApp = acceptOverride $ \req f -> f $ responseLBS status200
     [("Accept", fromMaybe "" $ lookup "Accept" $ requestHeaders req)] ""
 
 caseAcceptOverride :: Assertion
-caseAcceptOverride = flip runSession aoApp $ do
+caseAcceptOverride = withSession aoApp $ do
     sres1 <- request defaultRequest
                 { queryString = []
                 , requestHeaders = [("Accept", "foo")]
@@ -329,13 +512,13 @@ caseAcceptOverride = flip runSession aoApp $ do
 
 caseDebugRequestBody :: Assertion
 caseDebugRequestBody = do
-    flip runSession (debugApp postOutput) $ do
+    withSession (debugApp postOutput) $ do
         let req = toRequest "application/x-www-form-urlencoded" "foo=bar&baz=bin"
         res <- srequest req
         assertStatus 200 res
 
     let qs = "?foo=bar&baz=bin"
-    flip runSession (debugApp $ getOutput params) $ do
+    withSession (debugApp $ getOutput params) $ do
         assertStatus 200 =<< request defaultRequest
                 { requestMethod = "GET"
                 , queryString = map (\(k,v) -> (k, Just v)) params
@@ -437,7 +620,7 @@ streamFileApp :: Application
 streamFileApp = streamFile $ \_ f -> f $ responseFile status200 [] testFile Nothing
 
 caseStreamFile :: Assertion
-caseStreamFile = flip runSession streamFileApp $ do
+caseStreamFile = withSession streamFileApp $ do
     sres <- request defaultRequest
     assertStatus 200 sres
     assertBodyContains "caseStreamFile" sres
@@ -449,7 +632,7 @@ streamLBSApp = streamFile $ \_ f -> f $ responseLBS status200
     "test"
 
 caseStreamLBS :: Assertion
-caseStreamLBS = flip runSession streamLBSApp $ do
+caseStreamLBS = withSession streamLBSApp $ do
     sres <- request defaultRequest
     assertStatus 200 sres
     assertBody "test" sres
@@ -466,7 +649,7 @@ caseModifyPostParamsInLogs = do
     testLogs formatRedacted outputRedacted
   where
     testLogs :: OutputFormat -> [(String, String)] -> Assertion
-    testLogs format output = flip runSession (debugApp format output) $ do
+    testLogs format output = withSession (debugApp format output) $ do
         let req = toRequest "application/x-www-form-urlencoded" "username=some_user&password=dont_show_me"
         res <- srequest req
         assertStatus 200 res
@@ -504,7 +687,7 @@ caseFilterRequestsInLogs = do
     testLogs formatFiltered pathHidden False
   where
     testLogs :: OutputFormat -> S8.ByteString -> Bool -> Assertion
-    testLogs format rpath haslogs = flip runSession (debugApp format rpath haslogs) $ do
+    testLogs format rpath haslogs = withSession (debugApp format rpath haslogs) $ do
         let req = flip SRequest "" $ setPath defaultRequest rpath
         res <- srequest req
         assertStatus 200 res
@@ -527,3 +710,28 @@ caseFilterRequestsInLogs = do
             actual `shouldBe` ""
 
         return res
+
+-- | Unit test to make sure 'parseQValueList' works correctly
+caseQValues :: Assertion
+caseQValues = do
+    let encodings = mconcat
+            -- This has weird white space on purpose, because this
+            -- should be acceptable according to RFC 7231
+            [ "deflate,   compress; q=0.813  ,gzip ;  q=0.9, * ;q=0, "
+            , "notq;charset=bar, "
+            , "noq;q=,   "
+            , "toolong;q=0.0023, "
+            , "toohigh ;q=1.1"
+            ]
+        qList = parseQValueList encodings
+        expected =
+            [ ("deflate", Just 1000)
+            , ("compress", Just 813)
+            , ("gzip", Just 900)
+            , ("*", Just 0)
+            , ("notq", Nothing)
+            , ("noq", Nothing)
+            , ("toolong", Nothing)
+            , ("toohigh", Nothing)
+            ]
+    assertEqual "invalid Q values" expected qList